refactor: insert, update 기능을 클래스로 추출한다.
